version: 2.8.0
Updates the Kingfisher dependency to version 8.5.0 and introduces a new PIPVideoCell with improved video player functionalities.
The new PIPVideoCell is implemented with SwiftUI and includes features like drag gesture for repositioning, full-screen support and orientation change handling.
Also replaces the old PIPVideoCell with the new one and improves the VideoMessageView.